Modularizando
se modulariza todas las carpetas del proyecto heroes
