+++ src/app/heroes/componentes/hero-search/herosearch.component.ts
@@ -6,8 +6,8 @@ import {
    debounceTime, distinctUntilChanged, switchMap
  } from 'rxjs/operators';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../../interfaces/hero';
+import { HeroService } from '../../service/hero.service';
 
 @Component({
   selector: 'app-hero-search',
