+++ src/app/heroes/componentes/messages/messages.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { MessageService } from '../message.service';
+import { MessageService } from '../../service/message.service';
 
 @Component({
   selector: 'app-messages',
