+++ src/app/heroes/interfaces/hero.ts
