+++ src/app/heroes/pages/heroes.component.html
