+++ src/app/heroes/pages/heroes.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../interfaces/hero';
+import { HeroService } from '../service/hero.service';
 
 @Component({
   selector: 'app-heroes',
