+++ src/app/heroes/service/message.service.spec.ts
@@ -1,7 +1,7 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async, inject } from '@angular/core/testing';
-import { MessageService } from '../message.service';
+import { MessageService } from './message.service';
 
 describe('Service: Message', () => {
   beforeEach(() => {
