+++ src/app/main/app-routing.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { HeroesComponent } from './heroes/heroes.component';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
+import { DashboardComponent } from '../heroes/componentes/dashboard/dashboard.component';
+import { HeroesComponent } from '../heroes/pages/heroes.component';
+import { HeroDetailComponent } from '../heroes/componentes/hero-detail/hero-detail.component';
 
 const routes: Routes = [
   { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
