+++ src/app/main/app.module.ts
@@ -4,14 +4,14 @@ import { FormsModule } from '@angular/forms'; // <-- NgModel lives here
 import { HttpClientModule } from '@angular/common/http';
 
 import { HttpClientInMemoryWebApiModule } from 'angular-in-memory-web-api';
-import { InMemoryDataService } from './In-memory-data.service';
+import { InMemoryDataService } from '../heroes/service/In-memory-data.service';
 
 import { AppComponent } from './componente/app.component';
-import { HeroesComponent } from './heroes/heroes.component';
-import { HeroDetailComponent } from './hero-detail/hero-detail.component';
-import { MessagesComponent } from './messages/messages.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { HeroSearchComponent } from './hero-search/herosearch.component';
+import { HeroesComponent } from '../heroes/pages/heroes.component';
+import { HeroDetailComponent } from '../heroes/componentes/hero-detail/hero-detail.component';
+import { MessagesComponent } from '../heroes/componentes/messages/messages.component';
+import { DashboardComponent } from '../heroes/componentes/dashboard/dashboard.component';
+import { HeroSearchComponent } from '../heroes/componentes/hero-search/herosearch.component';
 
 
 import { AppRoutingModule } from './app-routing.module'
