+++ src/app/main/componente/app.component.scss
