+++ src/main.ts
@@ -1,6 +1,6 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { AppModule } from './app/main/app.module';
 
 
 platformBrowserDynamic().bootstrapModule(AppModule)
